started breaking apart fragmentor into submodules + remembered to commit the state map directory (whoops)
