+++ file_shredder/lib/file_shredder/fragmentor.ex
@@ -25,17 +25,6 @@ defmodule FileShredder.Fragmentor do
     %{"payload" => chunk}
   end
 
-  defp dummy(chunk_size) do
-    if @debug do IO.puts("dummy...") end
-    chunk = to_string(:string.chars(0, chunk_size-1))# |> Utils.Crypto.pad(chunk_size)
-    %{"payload" => chunk}
-  end
-
-  defp gen_dummies(0, _chunk_size), do: []
-  defp gen_dummies(dummy_count, chunk_size) do
-    for _ <- 0..dummy_count-1, do: dummy(chunk_size)
-  end
-
   def fragment(file_path, n, password) when n > 1 do
     hashkey = Utils.Crypto.gen_key(password)
     file_name = Path.basename(file_path)
@@ -45,7 +34,6 @@ defmodule FileShredder.Fragmentor do
     padding = (n * (chunk_size - 1)) - file_size
     partial_pad = rem(padding, (chunk_size - 1))
     dummy_count = div((padding - partial_pad), (chunk_size - 1))
-
     
     {:ok, file_info_pid} = State.Map.start_link(
       %{
@@ -58,7 +46,7 @@ defmodule FileShredder.Fragmentor do
     frag_paths = file_path
     |> File.stream!([], chunk_size - 1)
     |> Stream.map(&pad_frag(&1, chunk_size)) # pad frags
-    |> Stream.concat(gen_dummies(dummy_count, chunk_size)) # add dummy frags
+    |> FileShredder.Fragmentor.Dummies.generate(dummy_count, chunk_size)
     |> Stream.with_index() # add sequence ID
     #|> Enum.map(&finish_frag(&1, file_info_pid))
     |> Utils.Parallel.pooled_map(&finish_frag(&1, file_info_pid))
